++weapon added

+++ RPG/Services/CharacterService/CharacterService.cs
@@ -1,7 +1,5 @@
 ﻿global using AutoMapper;
-using RPG.Data;
-using RPG.Dtos.Character;
-using RPG.Models;
+using System.Security.Claims;
 
 namespace RPG.Services.CharacterService
 {
@@ -9,10 +7,12 @@ namespace RPG.Services.CharacterService
     {
 
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
         private readonly IMapper _mapper;
-
-        public CharacterService(IMapper mapper, DataContext context)
+        
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
+            _httpContextAccessor = httpContextAccessor;
             _mapper = mapper;
             _context = context;
         }
@@ -21,9 +21,12 @@ namespace RPG.Services.CharacterService
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             var character = _mapper.Map<Character>(Newcharacter);
+            character.User = await _context.Users!.FirstOrDefaultAsync(u => u.Id == GetUserId());
             _context.Characters.Add(character);
             await _context.SaveChangesAsync();
-            serviceResponse.Data =await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+            serviceResponse.Data =await _context.Characters
+                .Where(c =>c.User!.Id == GetUserId())
+                .Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
             return serviceResponse;
         }
 
@@ -33,13 +36,15 @@ namespace RPG.Services.CharacterService
 
             try
             {
-                var character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
-                if (character is null)
+                var character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+                if (character is null || character.User.Id != GetUserId())
                 {
                     throw new Exception($"Character with ID '{id}' not found");
                 }
                 _context.Characters.Remove(character); 
-                serviceResponse.Data =await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+                serviceResponse.Data =await _context.Characters
+                    .Where(c => c.User.Id == GetUserId())
+                    .Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
                await _context.SaveChangesAsync();
             }
             catch (Exception ex)
@@ -56,7 +61,7 @@ namespace RPG.Services.CharacterService
             try
             {
                 var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-                var dbCharacter = await _context.Characters.ToListAsync();
+                var dbCharacter = await _context.Characters.Where(c => c.User!.Id == GetUserId()).ToListAsync();
                 serviceResponse.Data = dbCharacter.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
                 return serviceResponse;
 
@@ -74,19 +79,22 @@ namespace RPG.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            var dbCharacter = await _context.Characters
+                .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
             return serviceResponse;
         }
-
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext!.User.FindFirstValue(ClaimTypes.NameIdentifier)!);
         public async Task<ServiceResponse<GetCharacterDto>> UpdateCharacter(UpdateCharacterDto updatedCharacter)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
 
             try
             {
-                var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
-                if(dbCharacter is null)
+                var dbCharacter = await _context.Characters
+                    .Include(c=> c.User)
+                    .FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+                if(dbCharacter is null || dbCharacter.User.Id != GetUserId())
                 {
                     throw new Exception($"Character with ID '{updatedCharacter.Id}' not found");
                 }
